[MISC] * updated setup.sh to set some environment variables in python virtual env * renamed extract validation set to extract dataset (more fitting) * removed old model directory, placed it in dnn directory * fixed some issues with eval_evaluation.py

+++ setup.sh
@@ -1,11 +1,56 @@
 #!/bin/sh
 
+[ $# -eq 0 ] && log 'Usage: ./setup.sh ([path to ember training dataset] [path to ember training labels] | --set-variables)' && exit 1
+
+log()
+{
+    echo "[SETUP] $1"
+}
+
+set_env()
+{
+    PATH_TO_PTH="$PWD/env/lib/python3.6/site-packages/set_env.pth"
+    >"$PATH_TO_PTH"
+    echo -n "import os; " >> "$PATH_TO_PTH"
+    echo -n "os.environ['ESLR_PROJECT_SOURCE_DIR']='$PWD'; " >> "$PATH_TO_PTH"
+    echo -n "os.environ['TRAINING_DATA']='$PWD/dataset/Xtraining.dat'; " >> "$PATH_TO_PTH"
+    echo -n "os.environ['VALIDATION_DATA']='$PWD/dataset/Xvalidation.dat'; " >> "$PATH_TO_PTH"
+    echo -n "os.environ['TRAINING_LABELS']='$PWD/dataset/Ytraining.dat'; " >> "$PATH_TO_PTH"
+    echo -n "os.environ['VALIDATION_LABELS']='$PWD/dataset/Yvalidation.dat'; " >> "$PATH_TO_PTH"
+    echo -n "os.environ['DNN_MODEL']='$PWD/src/dnn/models/04/04'; " >> "$PATH_TO_PTH"
+}
+
+SCRIPT_DIR="$(dirname $0)"
+cd "$SCRIPT_DIR"
+
+if [ "$1" = "--set-variables" ]; then
+    log 'Skipping dataset creation'
+    set_env
+    log 'Setted env variables'
+    exit 0
+fi
+
+# create virtual env
 rm -rf env/ 2>/dev/null
+python3 -m venv --python=/usr/bin/python3.6 env
+log '(1/4) Created virtual env'
 
-python3 -m venv env
-
+# install requirements
 cd env
 . bin/activate
 cd ..
-
 pip3 install -r requirements.txt
+log '(2/4) Installed requirements'
+
+# set environment variables
+set_env
+log '(3/4) Set environment variables'
+
+
+# fill environment variables with actual files
+PATH_TO_DATASET="$1"
+PATH_TO_EXTRACT_DATASET_SCRIPT='utils/extract_datasets.py'
+python3 "$PATH_TO_EXTRACT_DATASET_SCRIPT" "$1" "$2"
+log '(4/4) Finished dataset extraction'
+
+log 'Setup completed'
